+++ packages/stream_video_flutter/lib/src/theme/components/components.dart
@@ -3,6 +3,7 @@ export 'call_button_badge_theme.dart';
 export 'call_control_bar_theme.dart';
 export 'call_duration_badge_theme.dart';
 export 'call_participants_grid_theme.dart';
+export 'call_participants_spotlight_theme.dart';
 export 'connection_quality_indicator_theme.dart';
 export 'desktop_screen_share_selector_theme.dart';
 export 'floating_participant_tile_theme.dart';